+++ rumq-core/src/mqtt4/asyncdeserialize.rs
@@ -1,24 +1,19 @@
-use crate::*;
-use byteorder::ReadBytesExt;
-use std::io::Read;
+use crate::mqtt4::*;
+use async_trait::async_trait;
+use tokio::io::AsyncReadExt;
 
-pub trait MqttRead: ReadBytesExt {
-    fn mqtt_read(&mut self) -> Result<Packet, Error> {
-        let packet_type = self.read_u8()?;
-        let remaining_len = self.read_remaining_length()?;
+#[async_trait]
+pub trait AsyncMqttRead: AsyncReadExt + Unpin {
+    async fn async_mqtt_read(&mut self) -> Result<Packet, Error> {
+        let packet_type = self.read_u8().await?;
+        let remaining_len = self.read_remaining_length().await?;
 
-        self.deserialize(packet_type, remaining_len)
+        self.async_deserialize(packet_type, remaining_len).await
     }
 
-    fn read_packet_type_and_remaining_length(&mut self) -> Result<(u8, usize), Error> {
-        let packet_type = self.read_u8()?;
-        let remaining_len = self.read_remaining_length()?;
-        Ok((packet_type, remaining_len))
-    }
-
-    fn deserialize(&mut self, byte1: u8, remaining_len: usize) -> Result<Packet, Error> {
+    async fn async_deserialize(&mut self, byte1: u8, remaining_len: usize) -> Result<Packet, Error> {
         let kind = packet_type(byte1 >> 4)?;
-
+        
         if remaining_len == 0 {
             // no payload packets
             return match kind {
@@ -30,35 +25,39 @@ pub trait MqttRead: ReadBytesExt {
         }
 
         match kind {
-            PacketType::Connect => Ok(Packet::Connect(self.read_connect()?)),
-            PacketType::Connack => Ok(Packet::Connack(self.read_connack(remaining_len)?)),
-            PacketType::Publish => Ok(Packet::Publish(self.read_publish(byte1, remaining_len)?)),
+            PacketType::Connect => Ok(Packet::Connect(self.read_connect().await?)),
+            PacketType::Connack => Ok(Packet::Connack(self.read_connack(remaining_len).await?)),
+            PacketType::Publish => Ok(Packet::Publish(self.read_publish(byte1, remaining_len).await?)),
             PacketType::Puback if remaining_len != 2 => Err(Error::PayloadSizeIncorrect),
             PacketType::Puback => {
-                let pkid = self.read_u16::<byteorder::BigEndian>()?;
+                let pkid = self.read_u16().await?;
                 Ok(Packet::Puback(PacketIdentifier(pkid)))
             }
             PacketType::Pubrec if remaining_len != 2 => Err(Error::PayloadSizeIncorrect),
             PacketType::Pubrec => {
-                let pkid = self.read_u16::<byteorder::BigEndian>()?;
+                let pkid = self.read_u16().await?;
                 Ok(Packet::Pubrec(PacketIdentifier(pkid)))
             }
             PacketType::Pubrel if remaining_len != 2 => Err(Error::PayloadSizeIncorrect),
             PacketType::Pubrel => {
-                let pkid = self.read_u16::<byteorder::BigEndian>()?;
+                let pkid = self.read_u16().await?;
                 Ok(Packet::Pubrel(PacketIdentifier(pkid)))
             }
             PacketType::Pubcomp if remaining_len != 2 => Err(Error::PayloadSizeIncorrect),
             PacketType::Pubcomp => {
-                let pkid = self.read_u16::<byteorder::BigEndian>()?;
+                let pkid = self.read_u16().await?;
                 Ok(Packet::Pubcomp(PacketIdentifier(pkid)))
             }
-            PacketType::Subscribe => Ok(Packet::Subscribe(self.read_subscribe(remaining_len)?)),
-            PacketType::Suback => Ok(Packet::Suback(self.read_suback(remaining_len)?)),
-            PacketType::Unsubscribe => Ok(Packet::Unsubscribe(self.read_unsubscribe(remaining_len)?)),
+            PacketType::Subscribe => Ok(Packet::Subscribe(
+                self.read_subscribe(remaining_len).await?,
+            )),
+            PacketType::Suback => Ok(Packet::Suback(self.read_suback(remaining_len).await?)),
+            PacketType::Unsubscribe => Ok(Packet::Unsubscribe(
+                self.read_unsubscribe(remaining_len).await?,
+            )),
             PacketType::Unsuback if remaining_len != 2 => Err(Error::PayloadSizeIncorrect),
             PacketType::Unsuback => {
-                let pkid = self.read_u16::<byteorder::BigEndian>()?;
+                let pkid = self.read_u16().await?;
                 Ok(Packet::Unsuback(PacketIdentifier(pkid)))
             }
             PacketType::Pingreq => Err(Error::IncorrectPacketFormat),
@@ -67,10 +66,10 @@ pub trait MqttRead: ReadBytesExt {
         }
     }
 
-    fn read_connect(&mut self) -> Result<Connect, Error> {
-        let protocol_name = self.read_mqtt_string()?;
-        let protocol_level = self.read_u8()?;
-
+    async fn read_connect(&mut self) -> Result<Connect, Error> {
+        let protocol_name = self.read_mqtt_string().await?;
+        let protocol_level = self.read_u8().await?;
+        
         if protocol_name != "MQTT" {
             return Err(Error::InvalidProtocolLevel(protocol_name, protocol_level));
         }
@@ -80,9 +79,9 @@ pub trait MqttRead: ReadBytesExt {
             _ => return Err(Error::InvalidProtocolLevel(protocol_name, protocol_level)),
         };
 
-        let connect_flags = self.read_u8()?;
-        let keep_alive = self.read_u16::<byteorder::BigEndian>()?;
-        let client_id = self.read_mqtt_string()?;
+        let connect_flags = self.read_u8().await?;
+        let keep_alive = self.read_u16().await?;
+        let client_id = self.read_mqtt_string().await?;
 
         let last_will = match connect_flags & 0b100 {
             0 => {
@@ -92,8 +91,8 @@ pub trait MqttRead: ReadBytesExt {
                 None
             }
             _ => {
-                let will_topic = self.read_mqtt_string()?;
-                let will_message = self.read_mqtt_string()?;
+                let will_topic = self.read_mqtt_string().await?;
+                let will_message = self.read_mqtt_string().await?;
                 let will_qos = qos((connect_flags & 0b11000) >> 3)?;
                 Some(LastWill {
                     topic: will_topic,
@@ -106,12 +105,12 @@ pub trait MqttRead: ReadBytesExt {
 
         let username = match connect_flags & 0b10000000 {
             0 => None,
-            _ => Some(self.read_mqtt_string()?),
+            _ => Some(self.read_mqtt_string().await?),
         };
 
         let password = match connect_flags & 0b01000000 {
             0 => None,
-            _ => Some(self.read_mqtt_string()?),
+            _ => Some(self.read_mqtt_string().await?),
         };
 
         Ok(Connect {
@@ -125,36 +124,36 @@ pub trait MqttRead: ReadBytesExt {
         })
     }
 
-    fn read_connack(&mut self, remaining_len: usize) -> Result<Connack, Error> {
+    async fn read_connack(&mut self, remaining_len: usize) -> Result<Connack, Error> {
         // check remaining number of bytes (ignoring fixed header length). fixed header
         if remaining_len != 2 {
             return Err(Error::PayloadSizeIncorrect);
         }
-        let flags = self.read_u8()?;
-        let return_code = self.read_u8()?;
+        let flags = self.read_u8().await?;
+        let return_code = self.read_u8().await?;
         Ok(Connack {
             session_present: (flags & 0x01) == 1,
             code: connect_return(return_code)?,
         })
     }
 
-    fn read_publish(&mut self, byte1: u8, remaining_len: usize) -> Result<Publish, Error> {
+    async fn read_publish(&mut self, byte1: u8, remaining_len: usize) -> Result<Publish, Error> {
         let qos = qos((byte1 & 0b0110) >> 1)?;
         let dup = (byte1 & 0b1000) != 0;
         let retain = (byte1 & 0b0001) != 0;
 
-        let topic_name = self.read_mqtt_string()?;
+        let topic_name = self.read_mqtt_string().await?;
 
         // Packet identifier exists where QoS > 0
         let pkid = match qos {
             QoS::AtMostOnce => None,
-            QoS::AtLeastOnce | QoS::ExactlyOnce => Some(PacketIdentifier(self.read_u16::<byteorder::BigEndian>()?)),
+            QoS::AtLeastOnce | QoS::ExactlyOnce => Some(PacketIdentifier(self.read_u16().await?))
         };
 
         // variable header len = len of topic (2 bytes) + topic.len() + [optional packet id (2 bytes)]
         let variable_header_len = match qos {
             QoS::AtMostOnce => 2 + topic_name.len(),
-            QoS::AtLeastOnce | QoS::ExactlyOnce => 2 + topic_name.len() + 2,
+            QoS::AtLeastOnce | QoS::ExactlyOnce => 2 + topic_name.len() + 2
         };
 
         let payload_len = remaining_len - variable_header_len;
@@ -162,7 +161,7 @@ pub trait MqttRead: ReadBytesExt {
         // read publish payload into the buffer
         let mut payload = Vec::with_capacity(payload_len);
         let mut s = self.take(payload_len as u64);
-        s.read_to_end(&mut payload)?;
+        s.read_to_end(&mut payload).await?;
         Ok(Publish {
             dup,
             qos,
@@ -173,8 +172,8 @@ pub trait MqttRead: ReadBytesExt {
         })
     }
 
-    fn read_subscribe(&mut self, remaining_len: usize) -> Result<Subscribe, Error> {
-        let pkid = self.read_u16::<byteorder::BigEndian>()?;
+    async fn read_subscribe(&mut self, remaining_len: usize) -> Result<Subscribe, Error> {
+        let pkid = self.read_u16().await?;
 
         // variable header size = 2
         // variable header + payload - variable header
@@ -182,8 +181,8 @@ pub trait MqttRead: ReadBytesExt {
         let mut topics = Vec::with_capacity(1);
 
         while payload_bytes > 0 {
-            let topic_filter = self.read_mqtt_string()?;
-            let requested_qos = self.read_u8()?;
+            let topic_filter = self.read_mqtt_string().await?;
+            let requested_qos = self.read_u8().await?;
             payload_bytes -= topic_filter.len() + 3;
             topics.push(SubscribeTopic {
                 topic_path: topic_filter,
@@ -197,13 +196,13 @@ pub trait MqttRead: ReadBytesExt {
         })
     }
 
-    fn read_suback(&mut self, remaining_len: usize) -> Result<Suback, Error> {
-        let pkid = self.read_u16::<byteorder::BigEndian>()?;
+    async fn read_suback(&mut self, remaining_len: usize) -> Result<Suback, Error> {
+        let pkid = self.read_u16().await?;
         let mut payload_bytes = remaining_len - 2;
         let mut return_codes = Vec::with_capacity(payload_bytes);
 
         while payload_bytes > 0 {
-            let return_code = self.read_u8()?;
+            let return_code = self.read_u8().await?;
             if return_code >> 7 == 1 {
                 return_codes.push(SubscribeReturnCodes::Failure)
             } else {
@@ -218,13 +217,13 @@ pub trait MqttRead: ReadBytesExt {
         })
     }
 
-    fn read_unsubscribe(&mut self, remaining_len: usize) -> Result<Unsubscribe, Error> {
-        let pkid = self.read_u16::<byteorder::BigEndian>()?;
+    async fn read_unsubscribe(&mut self, remaining_len: usize) -> Result<Unsubscribe, Error> {
+        let pkid = self.read_u16().await?;
         let mut payload_bytes = remaining_len - 2;
         let mut topics = Vec::with_capacity(1);
 
         while payload_bytes > 0 {
-            let topic_filter = self.read_mqtt_string()?;
+            let topic_filter = self.read_mqtt_string().await?;
             payload_bytes -= topic_filter.len() + 2;
             topics.push(topic_filter);
         }
@@ -235,35 +234,23 @@ pub trait MqttRead: ReadBytesExt {
         })
     }
 
-    fn read_mqtt_string(&mut self) -> Result<String, Error> {
-        let len = self.read_u16::<byteorder::BigEndian>()? as usize;
+    async fn read_mqtt_string(&mut self) -> Result<String, Error> {
+        let len = self.read_u16().await? as usize;
         let mut data = Vec::with_capacity(len);
-
+        
         let mut s = self.take(len as u64);
-        s.read_to_end(&mut data)?;
+        s.read_to_end(&mut data).await?;
 
         Ok(String::from_utf8(data)?)
     }
 
-    fn header_len(&self, remaining_len: usize) -> usize {
-        if remaining_len >= 2_097_152 {
-            4 + 1
-        } else if remaining_len >= 16_384 {
-            3 + 1
-        } else if remaining_len >= 128 {
-            2 + 1
-        } else {
-            1 + 1
-        }
-    }
-
-    fn read_remaining_length(&mut self) -> Result<usize, Error> {
+    async fn read_remaining_length(&mut self) -> Result<usize, Error> {
         let mut mult: usize = 1;
         let mut len: usize = 0;
         let mut done = false;
 
         while !done {
-            let byte = self.read_u8()? as usize;
+            let byte = self.read_u8().await? as usize;
             len += (byte & 0x7F) * mult;
             mult *= 0x80;
 
@@ -278,32 +265,35 @@ pub trait MqttRead: ReadBytesExt {
 }
 
 /// Implement MattRead for every AsyncReadExt type (and hence AsyncRead type)
-impl<R: ReadBytesExt + ?Sized> MqttRead for R {}
+impl<R: AsyncReadExt + ?Sized + Unpin> AsyncMqttRead for R {}
 
 #[cfg(test)]
 mod test {
-    use super::MqttRead;
-    use crate::{Connack, Connect, Packet, Publish, Suback, Subscribe, Unsubscribe};
-    use crate::{ConnectReturnCode, LastWill, PacketIdentifier, Protocol, QoS, SubscribeReturnCodes, SubscribeTopic};
+    use super::AsyncMqttRead;
+    use super::{Connack, Connect, Packet, Publish, Suback, Subscribe, Unsubscribe};
+    use super::{
+        ConnectReturnCode, LastWill, PacketIdentifier, Protocol, QoS, SubscribeReturnCodes,
+        SubscribeTopic,
+    };
     use std::io::Cursor;
 
-    #[test]
-    fn read_packet_connect_mqtt_protocol() {
+    #[tokio::test]
+    async fn read_packet_connect_mqtt_protocol() {
         let mut stream = Cursor::new(vec![
-            0x10, 39, // packet type, flags and remaining len
-            0x00, 0x04, 'M' as u8, 'Q' as u8, 'T' as u8, 'T' as u8, 0x04,       // variable header
-            0b11001110, // variable header. +username, +password, -will retain, will qos=1, +last_will, +clean_session
-            0x00, 0x0a, // variable header. keep alive = 10 sec
-            0x00, 0x04, 't' as u8, 'e' as u8, 's' as u8, 't' as u8, // payload. client_id
-            0x00, 0x02, '/' as u8, 'a' as u8, // payload. will topic = '/a'
-            0x00, 0x07, 'o' as u8, 'f' as u8, 'f' as u8, 'l' as u8, 'i' as u8, 'n' as u8,
-            'e' as u8, // payload. variable header. will msg = 'offline'
-            0x00, 0x04, 'r' as u8, 'u' as u8, 'm' as u8, 'q' as u8, // payload. username = 'rumq'
-            0x00, 0x02, 'm' as u8, 'q' as u8, // payload. password = 'mq'
-            0xDE, 0xAD, 0xBE, 0xEF, // extra packets in the stream
+            0x10, 39,                                                       // packet type, flags and remaining len
+            0x00, 0x04, 'M' as u8, 'Q' as u8, 'T' as u8, 'T' as u8, 0x04,   // variable header
+            0b11001110,                                                     // variable header. +username, +password, -will retain, will qos=1, +last_will, +clean_session
+            0x00, 0x0a,                                                     // variable header. keep alive = 10 sec
+            0x00, 0x04, 't' as u8, 'e' as u8, 's' as u8, 't' as u8,         // payload. client_id
+            0x00, 0x02, '/' as u8, 'a' as u8,                               // payload. will topic = '/a'
+            0x00, 0x07, 'o' as u8, 'f' as u8, 'f' as u8, 'l' as u8, 
+            'i' as u8, 'n' as u8, 'e' as u8,                                // payload. variable header. will msg = 'offline'
+            0x00, 0x04, 'r' as u8, 'u' as u8, 'm' as u8, 'q' as u8,         // payload. username = 'rumq'
+            0x00, 0x02, 'm' as u8, 'q' as u8,                               // payload. password = 'mq'
+            0xDE, 0xAD, 0xBE, 0xEF                                          // extra packets in the stream
         ]);
 
-        let packet = stream.mqtt_read().unwrap();
+        let packet = stream.async_mqtt_read().await.unwrap();
 
         assert_eq!(
             packet,
@@ -324,14 +314,14 @@ mod test {
         );
     }
 
-    #[test]
-    fn read_packet_connack_works() {
+    #[tokio::test]
+    async fn read_packet_connack_works() {
         let mut stream = Cursor::new(vec![
-            0b00100000, 0x02, // packet type, flags and remaining len
-            0x01, 0x00, // variable header. connack flags, connect return code
-            0xDE, 0xAD, 0xBE, 0xEF, // extra packets in the stream
+            0b00100000, 0x02,               // packet type, flags and remaining len
+            0x01, 0x00,                     // variable header. connack flags, connect return code
+            0xDE, 0xAD, 0xBE, 0xEF          // extra packets in the stream
         ]);
-        let packet = stream.mqtt_read().unwrap();
+        let packet = stream.async_mqtt_read().await.unwrap();
 
         assert_eq!(
             packet,
@@ -342,17 +332,17 @@ mod test {
         );
     }
 
-    #[test]
-    fn read_packet_publish_qos1_works() {
+    #[tokio::test]
+    async fn read_packet_publish_qos1_works() {
         let mut stream = Cursor::new(vec![
-            0b00110010, 11, // packet type, flags and remaining len
+            0b00110010, 11,                              // packet type, flags and remaining len
             0x00, 0x03, 'a' as u8, '/' as u8, 'b' as u8, // variable header. topic name = 'a/b'
-            0x00, 0x0a, // variable header. pkid = 10
-            0xF1, 0xF2, 0xF3, 0xF4, // publish payload
-            0xDE, 0xAD, 0xBE, 0xEF, // extra packets in the stream
+            0x00, 0x0a,                                  // variable header. pkid = 10
+            0xF1, 0xF2, 0xF3, 0xF4,                      // publish payload
+            0xDE, 0xAD, 0xBE, 0xEF                       // extra packets in the stream
         ]);
 
-        let packet = stream.mqtt_read().unwrap();
+        let packet = stream.async_mqtt_read().await.unwrap();
 
         assert_eq!(
             packet,
@@ -367,16 +357,16 @@ mod test {
         );
     }
 
-    #[test]
-    fn read_packet_publish_qos0_works() {
+    #[tokio::test]
+    async fn read_packet_publish_qos0_works() {
         let mut stream = Cursor::new(vec![
-            0b00110000, 7, // packet type, flags and remaining len
-            0x00, 0x03, 'a' as u8, '/' as u8, 'b' as u8, // variable header. topic name = 'a/b'
-            0x01, 0x02, // payload
-            0xDE, 0xAD, 0xBE, 0xEF, // extra packets in the stream
+            0b00110000, 7,                                  // packet type, flags and remaining len 
+            0x00, 0x03, 'a' as u8, '/' as u8, 'b' as u8,    // variable header. topic name = 'a/b'
+            0x01, 0x02,                                     // payload
+            0xDE, 0xAD, 0xBE, 0xEF                          // extra packets in the stream
         ]);
 
-        let packet = stream.mqtt_read().unwrap();
+        let packet = stream.async_mqtt_read().await.unwrap();
 
         assert_eq!(
             packet,
@@ -391,33 +381,33 @@ mod test {
         );
     }
 
-    #[test]
-    fn read_packet_puback_works() {
+    #[tokio::test]
+    async fn read_packet_puback_works() {
         let mut stream = Cursor::new(vec![
-            0b01000000, 0x02, // packet type, flags and remaining len
-            0x00, 0x0A, // fixed header. packet identifier = 10
-            0xDE, 0xAD, 0xBE, 0xEF, // extra packets in the stream
+            0b01000000, 0x02,                            // packet type, flags and remaining len 
+            0x00, 0x0A,                                  // fixed header. packet identifier = 10
+            0xDE, 0xAD, 0xBE, 0xEF                       // extra packets in the stream
         ]);
-        let packet = stream.mqtt_read().unwrap();
+        let packet = stream.async_mqtt_read().await.unwrap();
 
         assert_eq!(packet, Packet::Puback(PacketIdentifier(10)));
     }
 
-    #[test]
-    fn read_packet_subscribe_works() {
+    #[tokio::test]
+    async fn read_packet_subscribe_works() {
         let mut stream = Cursor::new(vec![
-            0b10000010, 20, // packet type, flags and remaining len
-            0x01, 0x04, // variable header. pkid = 260
-            0x00, 0x03, 'a' as u8, '/' as u8, '+' as u8, // payload. topic filter = 'a/+'
-            0x00,      // payload. qos = 0
-            0x00, 0x01, '#' as u8, // payload. topic filter = '#'
-            0x01,      // payload. qos = 1
-            0x00, 0x05, 'a' as u8, '/' as u8, 'b' as u8, '/' as u8, 'c' as u8, // payload. topic filter = 'a/b/c'
-            0x02,      // payload. qos = 2
-            0xDE, 0xAD, 0xBE, 0xEF, // extra packets in the stream
+            0b10000010, 20,                                                     // packet type, flags and remaining len
+            0x01, 0x04,                                                         // variable header. pkid = 260
+            0x00, 0x03, 'a' as u8, '/' as u8, '+' as u8,                        // payload. topic filter = 'a/+'
+            0x00,                                                               // payload. qos = 0
+            0x00, 0x01, '#' as u8,                                              // payload. topic filter = '#'
+            0x01,                                                               // payload. qos = 1
+            0x00, 0x05, 'a' as u8, '/' as u8, 'b' as u8, '/' as u8, 'c' as u8,  // payload. topic filter = 'a/b/c'
+            0x02,                                                               // payload. qos = 2
+            0xDE, 0xAD, 0xBE, 0xEF                                              // extra packets in the stream
         ]);
 
-        let packet = stream.mqtt_read().unwrap();
+        let packet = stream.async_mqtt_read().await.unwrap();
 
         assert_eq!(
             packet,
@@ -441,18 +431,18 @@ mod test {
         );
     }
 
-    #[test]
-    fn read_packet_unsubscribe_works() {
+    #[tokio::test]
+    async fn read_packet_unsubscribe_works() {
         let mut stream = Cursor::new(vec![
-            0b10100010, 17, // packet type, flags and remaining len
-            0x00, 0x0F, // variable header. pkid = 15
-            0x00, 0x03, 'a' as u8, '/' as u8, '+' as u8, // payload. topic filter = 'a/+'
-            0x00, 0x01, '#' as u8, // pyaload. topic filter = '#'
-            0x00, 0x05, 'a' as u8, '/' as u8, 'b' as u8, '/' as u8, 'c' as u8, // payload. topic filter = 'a/b/c'
-            0xDE, 0xAD, 0xBE, 0xEF, // extra packets in the stream
+            0b10100010, 17,                                                     // packet type, flags and remaining len 
+            0x00, 0x0F,                                                         // variable header. pkid = 15
+            0x00, 0x03, 'a' as u8, '/' as u8, '+' as u8,                        // payload. topic filter = 'a/+'
+            0x00, 0x01, '#' as u8,                                              // pyaload. topic filter = '#'
+            0x00, 0x05, 'a' as u8, '/' as u8, 'b' as u8, '/' as u8, 'c' as u8,  // payload. topic filter = 'a/b/c'
+            0xDE, 0xAD, 0xBE, 0xEF                                              // extra packets in the stream
         ]);
 
-        let packet = stream.mqtt_read().unwrap();
+        let packet = stream.async_mqtt_read().await.unwrap();
 
         assert_eq!(
             packet,
@@ -463,22 +453,25 @@ mod test {
         );
     }
 
-    #[test]
-    fn read_packet_suback_works() {
+    #[tokio::test]
+    async fn read_packet_suback_works() {
         let mut stream = Cursor::new(vec![
-            0x90, 4, // packet type, flags and remaining len
-            0x00, 0x0F, // variable header. pkid = 15
-            0x01, 0x80, // payload. return codes [success qos1, failure]
-            0xDE, 0xAD, 0xBE, 0xEF, // extra packets in the stream
+            0x90, 4,                    // packet type, flags and remaining len 
+            0x00, 0x0F,                 // variable header. pkid = 15
+            0x01, 0x80,                 // payload. return codes [success qos1, failure]
+            0xDE, 0xAD, 0xBE, 0xEF      // extra packets in the stream
         ]);
 
-        let packet = stream.mqtt_read().unwrap();
+        let packet = stream.async_mqtt_read().await.unwrap();
 
         assert_eq!(
             packet,
             Packet::Suback(Suback {
                 pkid: PacketIdentifier(15),
-                return_codes: vec![SubscribeReturnCodes::Success(QoS::AtLeastOnce), SubscribeReturnCodes::Failure]
+                return_codes: vec![
+                    SubscribeReturnCodes::Success(QoS::AtLeastOnce),
+                    SubscribeReturnCodes::Failure
+                ]
             })
         );
     }
