+++ rumq-core/src/mqtt4/topic.rs
